Move mercurial tools from bot unit tests in conftest
