+++ bot/tests/test_phabricator.py
@@ -7,11 +7,11 @@ import urllib.parse
 import responses
 
 from code_coverage_bot.phabricator import PhabricatorUploader
+from conftest import add_file
+from conftest import changesets
+from conftest import commit
+from conftest import copy_pushlog_database
 from conftest import covdir_report
-from mercurial import add_file
-from mercurial import changesets
-from mercurial import commit
-from mercurial import copy_pushlog_database
 
 
 @responses.activate
